reorganize proof of concept
