--- library/src/main/java/TreeData.java
@@ -1,9 +1,0 @@
-public class TreeData {
-  String key;
-  String text;
-
-  TreeData(String key, String text) {
-    this.key = key;
-    this.text = text;
-  }
-}
