+++ library/src/main/java/poc/YamlUtils.java
@@ -1,4 +1,4 @@
-/*
+package poc;/*
 Copyright 2020 Google LLC
 
 Licensed under the Apache License, Version 2.0 (the "License");
