--- library/src/test/java/treeUtilsTest.java
@@ -1,12 +1,0 @@
-// import org.junit.jupiter.api.BeforeEach;
-//
-// import static org.junit.jupiter.api.Assertions.*;
-//
-// class treeUtilsTest {
-//    TreeUtils treeUtils;
-//
-//    @BeforeEach
-//    void init(){
-//        treeUtils = new TreeUtils();
-//    }
-// }
