+++ library/src/test/resources/simplepetstore2.yaml
@@ -7,7 +7,7 @@ servers:
 paths:
   /pets:
     get:
-      summary: List all pets
+      summary: List every pet
       operationId: listPets
       tags:
         - pets
