+++ library/src/test/resources/simplepetstore3.yaml
@@ -1,0 +1,5 @@
+openapi: 3.0.0
+info:
+  title: Swagger Petstore
+  license:
+    name: MIT
